Implement ScrollViewManager
This improves the implementation for constraining horizontally
scrollable components to not be vertically scrollable. It introduces
the ScrollViewManager which handles the constraints plus the
positioning of headers and footer views inside of a component.
